These changes are to make gratings from spritetools follow the standard lab conventions. The following functions were incorrect:   singrat   cosgrat   hypergrat   polargrat   Make_2D_Sine   Make_2D_Cnc_Rdl   Make_2D_Hyperbolic
I have modified:
  singrat
  cosgrat
  hypergrat
  polargrat
to follow the gallant/mazer/schiller lab convention, which is as follows:
gratings are defined by their perpendicular angle (the "direction of motion")
relative to the 3:00 position. 3:00 is defined as 0 degrees. This means a
vertical grating has an orientation of 0 or 180 deg, horizaontal is 90 or 270
and a 45deg grating runs diagonally from 10:00 to 4:00-ish. Klar?

The Make_2D_* functions have been left alone, since they were directly
translated from Jack's original C code used to generate the stimuli used in
the original non-cartesian papers. These functions now print warnings when
called... it's probably best to not use them for anything in the future;
they're really only included as a historical reference point.

And, as Kate noted, the Sprite.rotate() method was doing clockwise rotations,
so instead of changing that and breaking existing code, I've added
Sprite.rotateCW() and Sprite.rotateCCW() methods to make the direction of
rotation explicit. The only rotate() method is the same as rotateCW(). But
most likely, people will want to use the rotateCCW() to keep to the standard
coordinate system. This should close the book on one or two long-standing
mysteries...

handmap.py has been changed to reflect these fixed. handmap (and
spotmap, for that matter) have always done the right thing, regardless of the
code -- they corrected this "bug" on the fly.

Finally, I know I've always told people to never do this, but I'm providing an
example filter program "samplefilter.py" that can be used as a model for
writing programs that filter raw pype datafiles. I used something like this to
"correct" the orientation data in a reasonably large # data files Jon
already collected that have the wrong orientation labels for the stimuli
due to this bug.. it can be done, just be careful, make sure you've
really thought about what you're doing AND KEEP BACKUPS!!!!

~~~~~~~~~~~~

The dacq.[ch] change is just an int -> void in the header for some
random function I fixed when I noticed a compile-time warning.. nothing
signficant. sorry to bundle that in here..



git-svn-id: svn+ssh://svn/auto/share/repos/LOCKDOWN/mlab-pype/pype2/trunk@146 97a921b0-85fb-0310-a969-e0731ffd8486